Generalize two functions
